feat: Move container class into own file
